update iresearch from the recent upstream

+++ 3rdParty/iresearch/core/utils/bit_utils.hpp
@@ -109,19 +109,19 @@ inline CONSTEXPR T ror( T value ) NOEXCEPT{
 #endif
 
 inline CONSTEXPR uint32_t zig_zag_encode32(int32_t v) NOEXCEPT {
-  return (v >> 31) ^ (uint32_t(v) << 1);
+  return (v >> 31) ^ (v << 1);
 }
 
 inline CONSTEXPR int32_t zig_zag_decode32(uint32_t v) NOEXCEPT {
-  return (v >> 1U) ^ -(v & 1U);
+  return (v >> 1) ^ -(v & 1);
 }
 
 inline CONSTEXPR uint64_t zig_zag_encode64(int64_t v) NOEXCEPT {
-  return (v >> 63) ^ (uint64_t(v) << 1);
+  return (v >> 63) ^ (v << 1);
 }
 
 inline CONSTEXPR int64_t zig_zag_decode64(uint64_t v) NOEXCEPT {
-  return (v >> 1U) ^ -(v & 1U);
+  return (v >> 1) ^ -(v & 1);
 }
 
 #if defined(_MSC_VER)
